+++ packages/rx_bloc_cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [5.2.0]
+* Add a script for running integration tests with Patrol
+* Updates generated project dependencies version
+
 ## [5.1.0]
 * Added Profile as an option under the flag `--enable-profile`
 * Update ruby version for generated `github` workflows